Corrected file name WelcomeScreen.js
